+++ src/constants/index.js
@@ -1,0 +1,3 @@
+export { default as CONFIG } from './config';
+export { MESSAGE, ERROR_MESSAGE } from './message';
+export { default as PRIZE } from './prize';
