+++ src/step1-index.js
@@ -1,6 +1,6 @@
-import CONFIG from './constants/config';
 import InputView from './views/InputView';
 import LottoController from './controllers/LottoController';
+import { CONFIG } from './constants';
 
 const app = {
   async play() {
