+++ src/validators/numberValidator.js
@@ -1,5 +1,4 @@
-import { ERROR_MESSAGE } from '../constants/message';
-import CONFIG from '../constants/config';
+import { CONFIG, ERROR_MESSAGE } from '../constants';
 
 const numberValidator = {
   validate(number) {
